Extract article-lead-asset
